+++ __tests__/fixtures.ts
@@ -1,6 +1,6 @@
 import fs from 'fs';
 
-import { Account, Provider, RpcProvider, ec, json } from '../src';
+import { Account, ProviderInterface, RpcProvider, SequencerProvider, ec, json } from '../src';
 import { CompiledContract } from '../src/types';
 
 const readContract = (name: string): CompiledContract =>
@@ -18,10 +18,20 @@ const DEFAULT_TEST_ACCOUNT_ADDRESS = // run `starknet-devnet --seed 0` and this 
 const DEFAULT_TEST_ACCOUNT_PRIVATE_KEY = '0xe3e70682c2094cac629f6fbed82c07cd';
 
 const BASE_URL = process.env.TEST_PROVIDER_BASE_URL || DEFAULT_TEST_PROVIDER_BASE_URL;
-export const IS_DEVNET = !BASE_URL.includes('starknet.io');
+const RPC_URL = process.env.TEST_RPC_URL;
+
+const IS_RPC = !!RPC_URL;
+const IS_RPC_DEVNET = Boolean(
+  RPC_URL && (RPC_URL.includes('localhost') || RPC_URL.includes('127.0.0.1'))
+);
+const IS_SEQUENCER = !IS_RPC;
+const IS_SEQUENCER_DEVNET = !BASE_URL.includes('starknet.io');
+export const IS_DEVNET = IS_SEQUENCER ? IS_SEQUENCER_DEVNET : IS_RPC_DEVNET;
 
 export const getTestProvider = () => {
-  const provider = new Provider({ sequencer: { baseUrl: BASE_URL } });
+  const provider = RPC_URL
+    ? new RpcProvider({ nodeUrl: RPC_URL })
+    : new SequencerProvider({ baseUrl: BASE_URL });
 
   if (IS_DEVNET) {
     // accelerate the tests when running locally
@@ -35,11 +45,11 @@ export const getTestProvider = () => {
 };
 
 // test account with fee token balance
-export const getTestAccount = (provider = getTestProvider(), isDevnet = IS_DEVNET) => {
+export const getTestAccount = (provider: ProviderInterface) => {
   let testAccountAddress = process.env.TEST_ACCOUNT_ADDRESS;
   let testAccountPrivateKey = process.env.TEST_ACCOUNT_PRIVATE_KEY;
 
-  if (!isDevnet) {
+  if (!IS_DEVNET) {
     if (!testAccountPrivateKey) {
       throw new Error('TEST_ACCOUNT_PRIVATE_KEY is not set');
     }
@@ -55,19 +65,7 @@ export const getTestAccount = (provider = getTestProvider(), isDevnet = IS_DEVNE
   return new Account(provider, testAccountAddress, ec.getKeyPair(testAccountPrivateKey));
 };
 
-export const getTestRpcProvider = () => {
-  const { TEST_RPC_URL } = process.env;
-
-  if (!TEST_RPC_URL) {
-    throw new Error('TEST_RPC_URL is not set');
-  }
-
-  return new RpcProvider({ nodeUrl: TEST_RPC_URL });
-};
-
-export const testIf = (condition: boolean) => (condition ? test : test.skip);
-export const describeIf = (condition: boolean) => (condition ? describe : describe.skip);
-export const testIfDevnet = testIf(IS_DEVNET);
-export const testIfNotDevnet = testIf(!IS_DEVNET);
-export const describeIfDevnet = describeIf(IS_DEVNET);
+const describeIf = (condition: boolean) => (condition ? describe : describe.skip);
+export const describeIfSequencer = describeIf(IS_DEVNET);
+export const describeIfRpc = describeIf(IS_RPC);
 export const describeIfNotDevnet = describeIf(!IS_DEVNET);